+++ src/main/java/com/lucendar/gnss/sdk/GnssConsts.java
@@ -1,0 +1,22 @@
+/*******************************************************************************
+ *  Copyright (c) 2019, 2022 lucendar.com.
+ *  All rights reserved.
+ *
+ *  Contributors:
+ *     KwanKin Yau (alphax@vip.163.com) - initial API and implementation
+ *******************************************************************************/
+package com.lucendar.gnss.sdk;
+
+public class GnssConsts {
+
+    public static final String DEFAULT_APP_ID = "";
+
+    public static class Features {
+        public static final String FEATURE__EXTERNAL_GATEWAY = "externalGateway";
+        public static final String FEATURE__ONLINE_TERMS_QRY_DISABLED = "onlineTermsQryDisabled";
+        /**
+         * A/V attribute query is supported.
+         */
+        public static final String FEATURE__AV_ATTRS_QRY = "avAttrsQry";
+    }
+}