+++ src/main/java/com/lucendar/gnss/sdk/strm/GnssOpenReplayStrmReq.java
@@ -5,7 +5,6 @@ import com.lucendar.strm.common.strm.AudioConfig;
 import com.lucendar.strm.common.strm.OpenStrmReq;
 import com.lucendar.strm.common.strm.ServerHint;
 
-
 import java.util.StringJoiner;
 
 public class GnssOpenReplayStrmReq implements GnssOpenStrmReq {
@@ -31,6 +30,7 @@ public class GnssOpenReplayStrmReq implements GnssOpenStrmReq {
 
     private String reqId;
     private boolean async = true;
+    private String appId;
     private String simNo;
     private byte channel;
     private byte mediaType;
@@ -52,7 +52,9 @@ public class GnssOpenReplayStrmReq implements GnssOpenStrmReq {
     private AudioConfig audioCfg;
     private OpenStrmReq.RtspSource rtspSrc;
     private String timedToken;
-
+    private Integer trace;
+    private String scheme;
+    private Boolean dontSendCloseIfInterrupt;
 
     @Override
     public String getReqId() {
@@ -75,6 +77,15 @@ public class GnssOpenReplayStrmReq implements GnssOpenStrmReq {
     @Override
     public boolean isLiveReq() {
         return false;
+    }
+
+    @Override
+    public String getAppId() {
+        return appId;
+    }
+
+    public void setAppId(String appId) {
+        this.appId = appId;
     }
 
     @Override
@@ -294,10 +305,38 @@ public class GnssOpenReplayStrmReq implements GnssOpenStrmReq {
     }
 
     @Override
+    public Integer getTrace() {
+        return trace;
+    }
+
+    public void setTrace(Integer trace) {
+        this.trace = trace;
+    }
+
+    @Override
+    public String getScheme() {
+        return scheme;
+    }
+
+    public void setScheme(String scheme) {
+        this.scheme = scheme;
+    }
+
+    @Override
+    public Boolean getDontSendCloseIfInterrupt() {
+        return dontSendCloseIfInterrupt;
+    }
+
+    public void setDontSendCloseIfInterrupt(Boolean dontSendCloseIfInterrupt) {
+        this.dontSendCloseIfInterrupt = dontSendCloseIfInterrupt;
+    }
+
+    @Override
     public String toString() {
         return new StringJoiner(", ", GnssOpenReplayStrmReq.class.getSimpleName() + "[", "]")
                 .add("reqId='" + reqId + "'")
                 .add("async=" + async)
+                .add("appId='" + appId + "'")
                 .add("simNo='" + simNo + "'")
                 .add("channel=" + channel)
                 .add("mediaType=" + mediaType)
@@ -319,6 +358,9 @@ public class GnssOpenReplayStrmReq implements GnssOpenStrmReq {
                 .add("audioCfg=" + audioCfg)
                 .add("rtspSrc=" + rtspSrc)
                 .add("timedToken='" + timedToken + "'")
+                .add("trace=" + trace)
+                .add("scheme='" + scheme + "'")
+                .add("dontSendCloseIfInterrupt=" + dontSendCloseIfInterrupt)
                 .toString();
     }
 }