+++ src/main/java/com/lucendar/gnss/sdk/strm/GnssOpenStrmReq.java
@@ -15,9 +15,40 @@ public interface GnssOpenStrmReq {
     String SUB_FORMAT__FMP4 = StreamingApi.STRM_SUB_FORMAT__FMP4;
     String SUB_FORMAT__MPEGTS = StreamingApi.STRM_SUB_FORMAT__MPEGTS;
 
+    /**
+     * DISABLED mode. DO NOT trace
+     */
+    int TRACE_MODE__DISABLED = 0;
+
+    /**
+     * SIMPLE mode. Trace:
+     * - stream state changes (StrmReq.state)
+     * - all terminal inbound/outbound message, including
+     */
+    int TRACE_MODE__SIMPLE = 1;
+
+    /**
+     * ADVANCED mode. Trace:
+     * - stream state changes (StrmReq.state)
+     * - stream media notification (StrmMediaNotif)
+     * - all terminal inbound/outbound message, including
+     */
+    int TRACE_MODE__ADVANCED = 2;
+
+    /**
+     * FULL mode. Trace:
+     * - stream state changes(StrmReq.state)
+     * - stream media notification (StrmMediaNotif)
+     * - stream data
+     * - all terminal inbound/outbound message, including
+     */
+    int TRACE_MODE__FULL = 3;
+
+
     String getReqId();
     boolean isAsync();
     boolean isLiveReq();
+    String getAppId();
     String getSimNo();
     byte getChannel();
     int getProto();
@@ -44,6 +75,7 @@ public interface GnssOpenStrmReq {
             return replayMediaType();
     }
 
+
     byte getCodeStrm();
     boolean isExclusive();
     boolean isSaveOnServer();
@@ -61,4 +93,62 @@ public interface GnssOpenStrmReq {
     OpenStrmReq.RtspSource getRtspSrc();
 
     String getTimedToken();
+
+    /**
+     * The trace mode should be applied to this request.
+     * Optional. `null` if not specified. Value domain: TRACE_MODE__XXXX constant defined in
+     * `com.lucendar.strm.common.strm.OpenStrmReq`.
+     *
+     * @return trace mode
+     */
+    Integer getTrace();
+
+    default int traceDef() {
+        Integer v = getTrace();
+        if (v != null)
+            return v;
+        else
+            return OpenStrmReq.TRACE_MODE__DISABLED;
+    }
+
+    default boolean isTraced() {
+        Integer trace = getTrace();
+        return trace != null && trace != OpenStrmReq.TRACE_MODE__DISABLED;
+    }
+
+    /**
+     * <pre>
+     * The scheme of media playing url. When the `proto` is `FLV` or `HLS`, the scheme of returned url in GNSS
+     * `strm/live/open`, `strm/replay/open` API is determined by the value of the `async` parameter:
+     * - when `async` is true, return URL with `wss` scheme.
+     * - when `async` is false, return URL with `https` scheme.
+     *
+     * If this logic is not appropriate, application can use this parameter to specify the scheme needed. At present,
+     * the following value(case-sensitive) are meaningful:
+     * - `http`: support FLV, HLS proto
+     * - `https`: support FLV, HLS proto
+     * - `ws`: support FLV, HLS proto
+     * - `wss`: support FLV, HLS proto
+     *
+     * This is a optional parameter, if supplied, the `scheme` must be supported by the `proto`.
+     *</pre>
+     * @return
+     */
+    String getScheme();
+
+    /**
+     * Whether prohibit the server send a stream close command to terminal if an interruption in the stream is detected.
+     *
+     * @return null let server determine what to do when the code stream is interrupted.
+     */
+    Boolean getDontSendCloseIfInterrupt();
+
+    default boolean dontSendCloseIfInterruptDef() {
+        Boolean v = getDontSendCloseIfInterrupt();
+        if (v != null)
+            return v;
+        else
+            return false;
+    }
+
 }